split globals into header and cpp file
